% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_speed.R
\name{adjust_speed}
\alias{adjust_speed}
\title{Estimate cumtime based on global speed}
\usage{
adjust_speed(gps_data, min_speed = 2, max_speed = 80, new_speed = NULL)
}
\arguments{
\item{gps_data}{A data.table with GPS-like rows.}

\item{min_speed}{Minimum speed to be considered as valid. Values below minimum
speed will be estimated as if they were NA. Default value is 2km/h.}

\item{max_speed}{Maximum speed to be considered as valid. Values above maximum
speed will be estimated as if they were NA. Default value is 80km/h.}

\item{new_speed}{Speed to replace missing values as well as values outside
min_speed and max_speed range. The default value is the average speed of the
gps data.}
}
\value{
A GPS data with updated values of cumtime.
}
\description{
Estimate cumtime of NA values of a GPS data based on the average
speed of the valid GPS rows.
}
\details{
This function estimates speed values using the average speed of the
whole GPS data, ignoring infinite, NA, and NAN values. Values above 80km/h and
below 2km/h are also ignored and their values will be updated with the average.
}
\examples{
poa <- read_gtfs(system.file("extdata/poa.zip", package="gtfs2gps"))

poa_gps <- gtfs2gps(poa)
poa_gps_new <- adjust_speed(poa_gps)
}
