% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cumtime.R
\name{estimate_cumtime}
\alias{estimate_cumtime}
\title{Estimate cumtime based on global speed}
\usage{
estimate_cumtime(gps_data)
}
\arguments{
\item{gps_data}{A data.table with GPS-like rows.}
}
\value{
A GPS data with updated values of cumtime.
}
\description{
Estimate cumtime of NA values of a GPS data based on the average
speed of the valid GPS rows.
}
\details{
This function estimates speed values using the average speed of the
whole GPS data, ignoring infinite, NA, and NAN values. Values above 80km/h and
below 2km/h are also ignored and their values will be updated with the average.
}
\examples{
poa <- read_gtfs(system.file("extdata/poa.zip", package="gtfs2gps"))

poa_gps <- gtfs2gps(poa)
poa_gps_new <- estimate_cumtime(poa_gps)
}
