% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify_gtfs.R
\name{filter_by_shape_id}
\alias{filter_by_shape_id}
\title{Filter GTFS by shape if}
\usage{
filter_by_shape_id(gtfs, shape_ids)
}
\arguments{
\item{gtfs}{A tibble in the tidytransit format.}

\item{shape_ids}{A vector of shape_ids belonging to the shapes of the GTFS data.}
}
\description{
Filter a GTFS data in the tidytransit format. It gets a set of
shape ids and returns a new GTFS by removing the other shape ids from the
shapes. It also removes the unnecessary trips, stop_times, stops, and routes
accordingly.
}
