% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps2sf_linestring.R
\name{gps2sf_linestring}
\alias{gps2sf_linestring}
\title{Converts a GPS-like data.table to a MultiLineString Simple Feature (sf) object}
\usage{
gps2sf_linestring(input_file, crs = 4326)
}
\arguments{
\item{input_file}{A data.table with timestamp data.}

\item{crs}{A Coordinate Reference System. The default value is 4326 (latlong WGS84).}
}
\value{
A simple feature (sf) object with MultiLineString data.
}
\description{
Every interval of GPS data points of constant speed for each trip_id is
 converted into a linestring segment.
}
\examples{
\donttest{

library(gtfs2gps)
library(dplyr)

poa <- read_gtfs(system.file("extdata/poa.zip", package = "gtfs2gps"))
poa_subset <- filter_by_shape_id(poa, c("T2-1", "A141-1")) \%>\%
  filter_single_trip()

poa_gps <- gtfs2gps(poa_subset)
poa_gps_sf <- gps2sf_linestring(poa_gps)
plot(poa_gps_sf$geometry)
}
}
