% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps_as_sf.R
\name{gps_as_sf}
\alias{gps_as_sf}
\title{Convert GPS data to a Simple Feature}
\usage{
gps_as_sf(gps, crs = 4326)
}
\arguments{
\item{gps}{A data.table with GPS data.}

\item{crs}{A Coordinate Reference System. The default value is 4326.}
}
\description{
Convert a GPS data stored in a data.table into a Simple Feature
}
