% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbsapply.R
\name{pbSapply}
\alias{pbSapply}
\title{Process a function in parallel}
\usage{
pbSapply(slaves, progress, X, FUN, ...)
}
\arguments{
\item{slaves}{Number of slaves.}

\item{progress}{Show progress bar}

\item{X}{A vector with the objects to be used as single arguments to FUN.}

\item{FUN}{A function to be called with each value of X.}

\item{...}{Other arguments passed to FUN after a value from X.}
}
\description{
Process a function in parallel using doSNOW.
This function is based on the function presented in
https://stackoverflow.com/questions/38199631/parsapply-and-progress-bar.
}
