% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_gtfs_freq.R
\name{test_gtfs_freq}
\alias{test_gtfs_freq}
\title{Test whether a GTFS feed is frequency based}
\usage{
test_gtfs_freq(gtfs)
}
\arguments{
\item{gtfs}{A GTFS data set stored in memory as a list of data.tables/data.frames.}
}
\value{
A string "frequency" or "simple".
}
\description{
Test whether a GTFS feed is frequency based or whether it
presents detailed time table for all routes and trip ids.
}
