% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cppSnapPoints}
\alias{cppSnapPoints}
\title{Snap points to closest stops}
\usage{
cppSnapPoints(data, ref)
}
\arguments{
\item{data}{A set of points to be snapped.}

\item{ref}{A set of points whose subset will be the snapped points, which are
the nearest points to the x values.}
}
\description{
Snap a set of points to the closest points available in a set of stops.
}
