% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gtfs.R
\name{filter_by_route_id}
\alias{filter_by_route_id}
\title{Filter GTFS data by route ids}
\usage{
filter_by_route_id(gtfs_data, route_ids)
}
\arguments{
\item{gtfs_data}{A list of data.tables read using gtfs2gps::reag_gtfs().}

\item{route_ids}{A vector of route_ids belonging to the routes and trips of the
gtfs_data data.}
}
\value{
A filtered GTFS data.
}
\description{
Filter a GTFS data by its route ids, subsetting routes
and trops. It also removes the
unnecessary shapes, trips, frequencies, and stops, accordingly.
}
\examples{
poa <- read_gtfs(system.file("extdata/poa.zip", package = "gtfs2gps"))

subset <- filter_by_route_id(poa, "T2")
}
