% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gtfs.R
\name{filter_by_route_type}
\alias{filter_by_route_type}
\title{Filter GTFS data by route type}
\usage{
filter_by_route_type(gtfs_data, route_types)
}
\arguments{
\item{gtfs_data}{A list of data.tables read using gtfs2gps::reag_gtfs().}

\item{route_types}{A vector of route_types belonging to the routes of the
gtfs_data data.}
}
\value{
A filtered GTFS data.
}
\description{
Filter a GTFS data by route type. It also removes the
unnecessary trips, shapes, stop_times, stops, and frequencies, accordingly.
}
\examples{
poa <- read_gtfs(system.file("extdata/poa.zip", package = "gtfs2gps"))

subset <- filter_by_route_type(poa, 3)
}
