% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gtfs.R
\name{filter_by_shape_id}
\alias{filter_by_shape_id}
\title{Filter GTFS data by shape ids}
\usage{
filter_by_shape_id(gtfs_data, shape_ids)
}
\arguments{
\item{gtfs_data}{A tibble in the tidytransit format.}

\item{shape_ids}{A vector of shape_ids belonging to the shapes of the gtfs_data data.}
}
\description{
Filter a GTFS data in the tidytransit format. It gets a set of
shape ids and returns a new gtfs_data by removing the other shape ids from the
shapes. It also removes the unnecessary trips, stop_times, stops, and routes
accordingly.
}
