% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_day_period.R
\name{filter_day_period}
\alias{filter_day_period}
\title{Filter GTFS data within a period of the day}
\usage{
filter_day_period(gps, start = "07:00", end = "09:00")
}
\arguments{
\item{start}{A string of type "hh:mm" indicating start of the period (defaults to "06:00")}

\item{end}{A string of type "hh:mm" indicating the end of the period (defaults to "09:00")}

\item{gtfs}{A GTFS data.}
}
\description{
Updates a GTFS feed filtering only the routes, shapes, trips, stops,
 agencies and services that are active within a given period of the day.
}
\examples{
\donttest{
library(gtfs2gps)

poa <- read_gtfs(system.file("extdata/poa.zip", package="gtfs2gps"))
poa_f <- filter_day_period(poa, start = "10:00", end = "19:00")
}
}
