% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gtfs.R
\name{filter_single_trip}
\alias{filter_single_trip}
\title{Filter GTFS trips in order to have one trip per shape_id}
\usage{
filter_single_trip(gtfs_data)
}
\arguments{
\item{gtfs_data}{A list of data.tables read using gtfs2gps::reag_gtfs().}
}
\description{
Filter a GTFS data by keeping only one trip per shape_id.
It also removes the unnecessary routes accordingly.
}
