% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gtfs.R
\name{filter_week_days}
\alias{filter_week_days}
\title{Filter GTFS trips operating on week days}
\usage{
filter_week_days(gtfs_data)
}
\arguments{
\item{gtfs_data}{A tibble in the tidytransit format.}
}
\description{
Filter a GTFS data read using gtfs2gps::read_gtfs(). It removes the
trips operating only saturday or sunday.
}
