% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs2gps.R
\name{gtfs2gps}
\alias{gtfs2gps}
\title{Convert GTFS to GPS given a spatial resolution}
\usage{
gtfs2gps(gtfs_data, filepath = NULL, spatial_resolution = 15,
  progress = TRUE)
}
\arguments{
\item{gtfs_data}{A path to a GTFS file to be converted to GPS, or a GTFS data
represented as a list of data.tables.}

\item{filepath}{Output file path. As default, the output is returned in R.
When this argument is set, each route is saved into a file within filepath,
with the name equals to its id. In this case, no output is returned.}

\item{spatial_resolution}{The spatial resolution in meters. Default is 15m.}

\item{progress}{Show a progress bar? Default is TRUE.}
}
\description{
Convert GTFS data to GPS format by sampling points using a
spatial resolution. This function creates additional points in order to
guarantee that two points in a same trip will have at most a given
distance, indicated as a spatial resolution.
}
