% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs2gps.R
\name{gtfs2gps}
\alias{gtfs2gps}
\title{Convert GTFS to GPS-like data given a spatial resolution}
\usage{
gtfs2gps(
  gtfs_data,
  spatial_resolution = 15,
  parallel = T,
  progress = TRUE,
  filepath = NULL,
  continue = FALSE
)
}
\arguments{
\item{gtfs_data}{A path to a GTFS file to be converted to GPS, or a GTFS data
represented as a list of data.tables.}

\item{spatial_resolution}{The spatial resolution in meters. Default is 15m.}

\item{parallel}{Decides whether the function should run in parallel. Defaults to TRUE.}

\item{progress}{Show a progress bar. Default is TRUE.}

\item{filepath}{Output file path. As default, the output is returned in R.
When this argument is set, each route is saved into a file within filepath,
with the name equals to its id. In this case, no output is returned.}

\item{continue}{Argument that can be used only with filepath. When TRUE, it
skips processing the shape identifiers that were already saved into files.
It is useful to continue processing a GTFS file that was stopped for some
reason. Default value is FALSE.}
}
\description{
Convert GTFS data to GPS format by sampling points using a
spatial resolution. This function creates additional points in order to
guarantee that two points in a same trip will have at most a given
distance, indicated as a spatial resolution.
}
\examples{
\donttest{
library(gtfs2gps)

poa <- gtfs2gps(system.file("extdata/poa.zip", package="gtfs2gps"))
}
}
