% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs2gps_single_dt.R
\name{gtfs2gps_dt_single}
\alias{gtfs2gps_dt_single}
\title{Convert GTFS to GPS given a spatial resolution}
\usage{
gtfs2gps_dt_single(gtfszip, filepath = ".", spatial_resolution = 15,
  week_days = TRUE)
}
\arguments{
\item{gtfszip}{A path to a GTFS file to be converted to GPS.}

\item{filepath}{Output file path.}

\item{spatial_resolution}{The spatial resolution in meters. Default is 15m.}

\item{week_days}{Use only the week days? Default is TRUE.}
}
\description{
Convert GTFS data to GPS format by sampling points using a
spatial resolution. This function creates additional points in order to
guarantee that two points in a same trip will have at most a given
distance, indicated as a spatial resolution.
}
