% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_as_sf.R
\name{gtfs_shapes_as_sf}
\alias{gtfs_shapes_as_sf}
\title{Convert GTFS shapes to simple feature object}
\usage{
gtfs_shapes_as_sf(gtfs, crs = 4326)
}
\arguments{
\item{gtfs}{A GTFS data.}

\item{crs}{The coordinate reference system represented as an EPSG code.
The default value is 4326 (latlong WGS84)}
}
\description{
Convert a GTFS shapes data loaded using gtfs2gps::read_gtf()
into a line simple feature (sf).
}
\examples{
\donttest{
poa <- read_gtfs(system.file("extdata/saopaulo.zip", package="gtfs2gps"))
poa_sf <- gtfs_shapes_as_sf(poa)
plot(sf::st_geometry(poa_sf), lwd = 2)
}
}
