% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gtfs.R
\name{read_gtfs}
\alias{read_gtfs}
\title{Read GTFS data into a list of data.tables}
\usage{
read_gtfs(gtfszip)
}
\arguments{
\item{gtfszip}{A zipped GTFS data.}
}
\description{
Read files of a zipped GTFS feed and load them to memory as a list of data.tables.
It will load the following files: "agency.txt", "calendar.txt", "routes.txt", "shapes.txt", 
"stop_times.txt", "stops.txt", and "trips.txt". If some of these files do not exit,
it will stop with an error.
}
\examples{
library(gtfs2gps)

poa <- read_gtfs(system.file("extdata/poa.zip", package="gtfs2gps"))
}
