% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_gtfs_freq.R
\name{test_gtfs_freq}
\alias{test_gtfs_freq}
\title{Test whether a GTFS feed is frequency based}
\usage{
test_gtfs_freq(gtfs)
}
\arguments{
\item{gtfs}{A GTFS data set stored in memory as a list of data.tables/data.frames.}
}
\description{
Test whether a GTFS feed is frequency based or whether the stop_times.txt file
presents detailed time table for all routes and trip ids .
}
\examples{
library(gtfs2gps)

# read a gtfs.zip to memory
poa <- read_gtfs(system.file("extdata/poa.zip", package="gtfs2gps"))

# Test whether a GTFS feed is frequency based
test_gtfs_freq(poa)

}
