% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gtfs.R
\name{write_gtfs}
\alias{write_gtfs}
\title{Write GTFS data into a zip file}
\usage{
write_gtfs(gtfs, file)
}
\arguments{
\item{gtfs}{A GTFS data.}

\item{file}{A .zip file name.}
}
\description{
Write GTFS stored in memory as a list of data.tables into a zipped GTFS feed.
}
\examples{
library(gtfs2gps)

poa <- read_gtfs(system.file("extdata/poa.zip", package="gtfs2gps"))
write_gtfs(poa, "mypoa.zip")
}
