% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_gtfs.R
\name{write_gtfs}
\alias{write_gtfs}
\title{Write GTFS data into a zip file}
\usage{
write_gtfs(gtfs, file)
}
\arguments{
\item{gtfs}{A GTFS data set stored in memory as a list of data.tables/data.frames.}

\item{file}{A .zip file name.}
}
\description{
Write GTFS stored in memory as a list of data.tables into a zipped GTFS feed.
This function overwrites the file zip file if it exists.
}
\examples{
library(gtfs2gps)

# read a gtfs.zip to memory
poa <- read_gtfs(system.file("extdata/poa.zip", package="gtfs2gps"))

# write GTFS data into a zip file
write_gtfs(poa, "mypoa.zip")

}
